% Generated by roxygen2 (4.1.1): do not edit by hand
% Please edit documentation in R/makeDecompMatrix.R
\name{makeDecompMatrix}
\alias{makeDecompMatrix}
\title{Create decomposition matrix}
\usage{
makeDecompMatrix(par, tauStr = "tau", transStr = "A", verbose = FALSE)
}
\arguments{
\item{par}{named array of turnover times (tau) and transfers (trans) following patterns specified in \code{publishedParameters.R}}

\item{tauStr}{string with the regexp that matches the name of the turnover times in par}

\item{transStr}{string with the regexp that matches the name of the transfer proportions in par}
}
\value{
matrix representing the decomposition rate
}
\description{
Create a decomposition matrix from specified turnover and transfers.
}
\examples{
makeDecompMatrix(unlist(list(tau1=10, tau2=100, A2=0.5)))
}

