% Generated by roxygen2 (4.1.1): do not edit by hand
% Please edit documentation in R/makeModelBounds.R
\name{makeModelBounds}
\alias{makeModelBounds}
\title{Create upper lower bounds for decomposition parameters}
\usage{
makeModelBounds(modelBase = publishedParameters(), verbose = FALSE)
}
\arguments{
\item{modelBase}{a list of model parameters. See \code{publishedParmaeters.R} for details on data structure.}

\item{verbose}{boolean flag to help debug}
}
\value{
a list of data frames defining the upper and lower bounds of parameters
}
\description{
Create upper lower bounds for decomposition parameters
}

