% Generated by roxygen2 (4.1.1): do not edit by hand
% Please edit documentation in R/runModel.R
\name{runModel}
\alias{runModel}
\title{Run a decomposition model}
\usage{
runModel(par, timeArr, cModel, verbose = FALSE)
}
\arguments{
\item{par}{a named numeric array. The allocation parameters end with 'aD' where D is an integer identifying the associated pool, and the pool name starts with 'lableN.' where lableN is the case lable. The turnover parameters are named 'turnoverD' where D is an integer identifying the associated pool. The transfer parameters are named 'transferM' where M is the index of the tranfer matrix that value is associated with.}

\item{timeArr}{numeric array defining time points we are interested in looking at.}

\item{cModel}{function that accepts the par parameters and define the carbon model.}

\item{verbose}{boolean flag for extra error messages}
}
\value{
a data.frame with the time and values for the carbon pools and CO2
}
\description{
Run a model across several different cases defined by different applocation parameters but which share the same decomposition and transfer parameters.
}

