% Generated by roxygen2 (4.1.1): do not edit by hand
% Please edit documentation in R/makeDecompMatrix.R
\name{makeDecompMatrix}
\alias{makeDecompMatrix}
\title{Create decomposition matrix}
\usage{
makeDecompMatrix(modelBase = publishedParameters())
}
\arguments{
\item{modelBase}{list of models following patterns specified in \code{publishedParameters.R}}
}
\value{
a list of matricies
}
\description{
Create a decomposition matrix from specified turnover and transfer arrays.
}

