% Generated by roxygen2 (4.1.1): do not edit by hand
% Please edit documentation in R/publishedParameters.R
\name{publishedParameters}
\alias{publishedParameters}
\title{Create list with published parameters}
\usage{
publishedParameters(nameArr = NULL)
}
\arguments{
\item{nameArr}{an array of characers naming the models to be pulled. Can be NULL if no subset is desired.}
}
\value{
A list of lists describing the parameters and citations for models. Decomposition pools are defined as pools which respire carbon dioxide and are not vegitation pools. Tau is in days and trans are fractional transfers to other carbon pools indexed as column matrix.
}
\description{
Create list with published parameters
}

